in_tail: rename tail_scan.c to tail_scan_glob.c
This is a preparation to add Windows support to tail_scan.c.

Since Windows does not support glob(3), we need to emulate the
functionality using Win32 API. For keeping the code base clean,
this patch moves the UNIX-only implementation to a separate file.

    1. tail_scan_glob.c  ... for UNIX platforms
    2. tail_scan_win32.c ... for Windows

Note: this commit does not contain any actual changes on the logic.
